Deleted EntityChevronTest from \\src\test\java, reovered old code from an old branch
