+++ src/test/java/test/entity/EntityChevronTest.java
@@ -8,6 +8,8 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.testng.annotations.Ignore;
 import runner.BaseTest;
+import org.openqa.selenium.*;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.annotations.Test;
 import org.testng.Assert;
 import runner.ProjectUtils;
@@ -17,9 +19,18 @@ import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
+import java.util.UUID;
 
 @Run(run = RunType.Multiple)
 public class EntityChevronTest extends BaseTest {
+    // remove me !!!!!
+    private WebDriverWait wait;
+
+    private static final String TITLE = UUID.randomUUID().toString();
+    private static final String INT_NUMBER = "123";
+    private static final String DOUBLE_NUMBER = "4.56";
+    private static final String STATUS_NEW = "Fulfillment";
+    private static final String STATUS_EDITED = "Pending";
 
     final String comments = "TEST1";
     final String int_ = "11";
@@ -76,7 +87,6 @@ public class EntityChevronTest extends BaseTest {
                 .sendKeys(title, int_, decimal, data, time)
                 .clickSaveButton();
         Assert.assertEquals(chevronPage.getRowCount(), rowCount + 1);
-
     }
 
     @Test(dependsOnMethods = "createNewRecord")
@@ -154,5 +164,79 @@ public class EntityChevronTest extends BaseTest {
         WebElement recheckFulfillment = driver.findElement(By.xpath("//a[@class = 'pa-chev-active']"));
         String ExpectedSign = "Fulfillment";
         Assert.assertEquals(ExpectedSign, recheckFulfillment.getText());
+    }
+
+    @Test(dependsOnMethods = "getFullSum")
+    private void addRecord() throws InterruptedException {
+
+        WebDriver driver = getDriver();
+
+        goChevronPage(driver);
+
+        ProjectUtils.click(driver, driver.findElement(By.xpath("//i[contains(text(), 'create_new_folder')]")));
+
+        ProjectUtils.click(driver, driver.findElement(By.xpath("//button[@data-id='string']")));
+        ProjectUtils.click(driver, driver.findElement(By.xpath(String.format("//span[contains(text(), '%s')]", STATUS_NEW))));
+        ProjectUtils.fill(getWait(), driver.findElement(By.id("text")), TITLE);
+        ProjectUtils.fill(getWait(), driver.findElement(By.id("int")), INT_NUMBER);
+        ProjectUtils.fill(getWait(), driver.findElement(By.id("decimal")), DOUBLE_NUMBER);
+        ProjectUtils.click(driver, driver.findElement(By.id("pa-entity-form-save-btn")));
+
+        WebElement row = findRow(driver);
+        Assert.assertNotNull(row, "New record hasn't been found in the list");
+    }
+
+    // remove me !!!!!
+    private WebDriverWait getWait() {
+        if (wait == null) {
+            wait = new WebDriverWait(getDriver(), 10);
+        }
+        return wait;
+    }
+
+
+    @Test(dependsOnMethods = "addRecord")
+    public void editStatus()  throws InterruptedException {
+
+        WebDriver driver = getDriver();
+
+        goChevronPage(driver);
+
+        WebElement row = findRow(driver);
+
+        ProjectUtils.click(driver, row.findElement(By.xpath(".//a[contains(@href, 'action_edit')]")));
+
+        ProjectUtils.click(driver, driver.findElement(By.xpath("//button[@data-id='string']")));
+        // click on status pending
+        ProjectUtils.click(driver, driver.findElement(By.xpath(String.format("//span[contains(text(), '%s')]", STATUS_EDITED))));
+        ProjectUtils.click(driver, driver.findElement(By.id("pa-entity-form-save-btn")));
+
+        ProjectUtils.click(driver, driver.findElement(By.xpath(String.format("//div[contains(@class,'card-body')]/div/a[contains(text(), '%s')]", STATUS_EDITED))));
+
+        WebElement rowEdited = findRow(driver);
+        Assert.assertNotNull(rowEdited, "Edited title hasn't been found in the filtered list");
+
+        ProjectUtils.click(driver, rowEdited.findElement(By.xpath(".//a[contains(@href, 'action_view')]")));
+        Assert.assertEquals(driver.findElement(By.xpath("//div[@id='crumbs']//a[@class='pa-chev-active']")).getText(), STATUS_EDITED, "New status is not equal");
+    }
+
+    private WebElement findRow(WebDriver driver) {
+
+        List<WebElement> rows = driver.findElements(By.xpath("//table[@id='pa-all-entities-table']/tbody/tr"));
+
+        for (WebElement row : rows) {
+            List<WebElement> cells = row.findElements(By.tagName("td"));
+            for (WebElement cell : cells) {
+                if (cell.getText().equals(TITLE)) {
+                    return row;
+                }
+            }
+        }
+        Assert.fail("Title hasn't been found in result table.");
+        return null;
+    }
+
+    private void goChevronPage(WebDriver driver) {
+        ProjectUtils.click(driver, driver.findElement(By.xpath("//p[contains(text(),'Chevron')]")));
     }
 }
